+++ scripts/storage_cleaner.py
@@ -27,11 +27,7 @@ from rich.progress import track
 
 from olmo import util
 from olmo.aliases import PathOrStr
-from olmo.checkpoint import (
-    Checkpointer,
-    LocalShardedCheckpointer,
-    TorchLegacyShardedCheckpointer,
-)
+from olmo.checkpoint import build_sharded_checkpointer
 from olmo.config import ShardedCheckpointerType, TrainConfig
 
 log = logging.getLogger(__name__)
@@ -854,7 +850,7 @@ def _add_training_config_to_checkpoint(local_checkpoint_dir: str, run_dir: str) 
 
 def _unshard_checkpoint(
     sharded_checkpoint_dir: str, dest_dir: str, run_dir: str, unsharding_config: UnshardCheckpointsConfig
-):
+) -> bool:
     local_storage = LocalFileSystemAdapter()
 
     # Download checkpoint to a temp dir if it is in cloud storage
@@ -890,13 +886,7 @@ def _unshard_checkpoint(
         # This is a hack, but decoupling unsharding for checkpoint saving/loading
         # seems like overkill.
         dummy_config = TrainConfig.new()
-        checkpointer: Checkpointer
-        if sharded_checkpoint_type == ShardedCheckpointerType.torch_legacy:
-            checkpointer = TorchLegacyShardedCheckpointer(dummy_config)
-        elif sharded_checkpoint_type == ShardedCheckpointerType.local:
-            checkpointer = LocalShardedCheckpointer(dummy_config)
-        else:
-            raise NotImplementedError(sharded_checkpoint_type)
+        checkpointer = build_sharded_checkpointer(dummy_config, name=sharded_checkpoint_type)
 
         model_state_dict, optim_state_dict, trainer_state_dict = checkpointer.unshard_checkpoint(
             sharding_input_dir
@@ -907,13 +897,14 @@ def _unshard_checkpoint(
             sharding_input_dir,
             sharding_output_dir,
             e,
+            exc_info=True,
         )
 
         if training_config_added:
             local_storage.delete_path(str(Path(sharding_input_dir) / CONFIG_YAML))
 
         local_storage.delete_path(sharding_output_dir)
-        return
+        return False
 
     # model
     model_output = str(Path(sharding_output_dir) / "model.pt")
@@ -945,6 +936,7 @@ def _unshard_checkpoint(
 
     dest_storage = _get_storage_adapter_for_path(dest_dir)
     dest_storage.upload(sharding_output_dir, dest_dir)
+    return True
 
 
 def _unshard_checkpoints(
@@ -990,11 +982,14 @@ def _unshard_checkpoints(
 
         if config.dry_run:
             log.info("Would unshard sharded checkpoint %s to %s", sharded_checkpoint_directory, dest_directory)
+            unsharding_suceeded = True
         else:
             log.info("Unsharding sharded checkpoint %s to %s", sharded_checkpoint_directory, dest_directory)
-            _unshard_checkpoint(sharded_checkpoint_directory, dest_directory, run_dir, config)
+            unsharding_suceeded = _unshard_checkpoint(
+                sharded_checkpoint_directory, dest_directory, run_dir, config
+            )
 
-        if config.delete_sharded_checkpoints:
+        if unsharding_suceeded and config.delete_sharded_checkpoints:
             assert run_dir == run_dir_or_archive
             if config.dry_run:
                 log.info("Would delete sharded checkpoint %s", sharded_checkpoint_directory)
@@ -1066,8 +1061,8 @@ def _get_wandb_config(wandb_run) -> TrainConfig:
     return wandb_config
 
 
-def _get_matching_wandb_runs(wandb_runs, training_run_dir: str) -> List:
-    config_path = os.path.join(training_run_dir, CONFIG_YAML)
+def _get_matching_wandb_runs(wandb_runs, checkpoint_dir: str) -> List:
+    config_path = os.path.join(checkpoint_dir, CONFIG_YAML)
     local_config_path = cached_path(config_path)
     train_config = TrainConfig.load(local_config_path)
 
@@ -1076,18 +1071,18 @@ def _get_matching_wandb_runs(wandb_runs, training_run_dir: str) -> List:
     ]
 
 
-def _get_wandb_path(run_dir: str) -> str:
+def _get_wandb_path(checkpoint_dir: str, run_dir: str) -> str:
     run_dir_storage = _get_storage_adapter_for_path(run_dir)
 
-    config_path = os.path.join(run_dir, CONFIG_YAML)
+    config_path = os.path.join(checkpoint_dir, CONFIG_YAML)
     if not run_dir_storage.is_file(config_path):
-        raise FileNotFoundError("No config file found in run dir, cannot get wandb path")
+        raise FileNotFoundError(f"No config file found in checkpoint dir {checkpoint_dir}, cannot get wandb path")
 
     local_config_path = cached_path(config_path)
     config = TrainConfig.load(local_config_path, validate_paths=False)
 
     if config.wandb is None or config.wandb.entity is None or config.wandb.project is None:
-        raise ValueError(f"Run at {run_dir} has missing wandb config, cannot get wandb run path")
+        raise ValueError(f"Checkpoint at {checkpoint_dir} has missing wandb config, cannot get wandb run path")
 
     wandb_runs = []
 
@@ -1096,18 +1091,17 @@ def _get_wandb_path(run_dir: str) -> str:
         wandb_runs += _get_wandb_runs_from_wandb_dir(run_dir_storage, wandb_dir, config)
 
     wandb_runs += _get_wandb_runs_from_train_config(config)
-
-    # Remove duplicate wandb runs based on run path, and wandb runs that do not match our run.
+    # Remove duplicate wandb runs based on run path, and wandb runs that do not match our checkpoint.
     wandb_runs = list({_get_wandb_path_from_run(wandb_run): wandb_run for wandb_run in wandb_runs}.values())
-    wandb_matching_runs = _get_matching_wandb_runs(wandb_runs, run_dir)
+    wandb_matching_runs = _get_matching_wandb_runs(wandb_runs, checkpoint_dir)
 
     if len(wandb_matching_runs) == 0:
-        raise RuntimeError(f"Failed to find any wandb runs for {run_dir}. Run might no longer exist")
+        raise RuntimeError(f"Failed to find any wandb runs for {checkpoint_dir}. Run might no longer exist")
 
     if len(wandb_matching_runs) > 1:
         wandb_run_urls = [wandb_run.url for wandb_run in wandb_matching_runs]
         raise RuntimeError(
-            f"Found {len(wandb_matching_runs)} runs matching run dir {run_dir}, cannot determine correct run: {wandb_run_urls}"
+            f"Found {len(wandb_matching_runs)} runs matching checkpoint dir {checkpoint_dir}, cannot determine correct run: {wandb_run_urls}"
         )
 
     return _get_wandb_path_from_run(wandb_matching_runs[0])
@@ -1203,14 +1197,15 @@ def _get_src_dest_pairs_for_copy(
 
     assert config.append_wandb_path and not is_archive_file
     checkpoint_to_wandb_path: Dict[str, str]
-    # TODO: Update _get_wandb_path to get the wandb path for a checkpoint rather than a run directory.
-    # A run directory could correspond to multiple wandb runs.
+    # No need to consider other checkpoints if we are filtering for a specific checkpoint
     if config.entry is not None and _is_checkpoint_dir(entry_path := os.path.join(run_dir, config.entry)):
         # No need to consider other checkpoints if we are filtering for a specific checkpoint
-        checkpoint_to_wandb_path = {entry_path: _get_wandb_path(run_dir)}
+        checkpoint_to_wandb_path = {entry_path: _get_wandb_path(entry_path, run_dir)}
     else:
         checkpoint_dirs = _get_checkpoint_dirs(run_dir, run_dir_storage)
-        checkpoint_to_wandb_path = {checkpoint_dir: _get_wandb_path(run_dir) for checkpoint_dir in checkpoint_dirs}
+        checkpoint_to_wandb_path = {
+            checkpoint_dir: _get_wandb_path(checkpoint_dir, run_dir) for checkpoint_dir in checkpoint_dirs
+        }
 
     src_dest_pairs: List[Tuple[str, str]] = []
     # Mappings of source checkpoint directories to destination checkpoint directories