+++ scripts/train.py
@@ -12,8 +12,15 @@ import torch.multiprocessing as mp
 import wandb
 from packaging import version
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
+from torch.distributed.fsdp import ShardingStrategy
+from torch.nn.parallel import DistributedDataParallel as DDP
 
-from olmo.config import CheckpointType, TrainConfig
+from olmo.config import (
+    CheckpointType,
+    DDPGradSyncMode,
+    DistributedStrategy,
+    TrainConfig,
+)
 from olmo.data import build_train_dataloader
 from olmo.eval import build_evaluators
 from olmo.exceptions import OLMoCliError, OLMoConfigurationError
@@ -24,12 +31,18 @@ from olmo.torch_util import (
     get_default_device,
     get_global_rank,
     get_local_rank,
+    get_local_world_size,
     get_world_size,
     peak_gpu_memory,
     seed_all,
 )
 from olmo.train import Trainer
-from olmo.util import clean_opt, log_extra_field, prepare_cli_environment
+from olmo.util import (
+    add_cached_path_clients,
+    clean_opt,
+    log_extra_field,
+    prepare_cli_environment,
+)
 
 log = logging.getLogger("train")
 
@@ -117,42 +130,90 @@ def main(cfg: TrainConfig) -> None:
     olmo_model = OLMo(cfg.model)
     log.info(f"Total number of parameters: {olmo_model.num_params():,d}")
     log.info(f"Number of non-embedding parameters: {olmo_model.num_params(include_embedding=False):,d}")
-    log.info(f"Peak GPU Memory (MB) before FSDP: {int(peak_gpu_memory() or 0)}")
+    log.info(f"Peak GPU Memory (MB) before {cfg.distributed_strategy}: {int(peak_gpu_memory() or 0)}")
 
     olmo_model.set_activation_checkpointing(cfg.activation_checkpointing)
 
-    # Wrap the model in FSDP.
-    log.info("Wrapping model with FDSP...")
-    wrap_policy = olmo_model.get_fsdp_wrap_policy(cfg.fsdp.wrapping_strategy)
+    if cfg.distributed_strategy == DistributedStrategy.ddp:
+        log.info("Wrapping model with DDP...")
+        assert cfg.ddp is not None, "DistributedStrategy ddp needs cfg.ddp to be set!"
 
-    if version.parse(torch.__version__) >= version.parse("2.1.0"):
-        # This prevents any parameters from being initialized twice
-        def dummy_init_fn(module: torch.nn.Module) -> None:
-            module.to_empty(device=get_default_device())
+        if cfg.model.init_device != "cuda":
+            raise OLMoConfigurationError("DDP does not work with init_device set to anything other than `cuda`.")
 
-        param_init_fn = dummy_init_fn
-    else:
+        if cfg.ddp.find_unused_params is True and cfg.ddp.grad_sync_mode != DDPGradSyncMode.micro_batch:
+            raise OLMoConfigurationError(
+                "`find_unused_params` is set to True. DDP needs to synchronize gradients for every micro-batch to avoid errors. Set `grad_sync_mode` to `micro_batch`."
+            )
+
         param_init_fn = None
-    fsdp_model = FSDP(
-        olmo_model,
-        sharding_strategy=cfg.fsdp.sharding_strategy,
-        mixed_precision=cfg.fsdp_precision,
-        auto_wrap_policy=wrap_policy,
-        use_orig_params=cfg.fsdp.use_orig_params,  # needed for compile and some of our optimizer/parameter metrics
-        limit_all_gathers=True,
-        device_id=get_local_rank(),
-        param_init_fn=param_init_fn,
-    )
+
+        # move to cuda before calling ddp
+        dist_model = DDP(olmo_model.to(device), find_unused_parameters=cfg.ddp.find_unused_params)
+    elif cfg.distributed_strategy == DistributedStrategy.fsdp:
+        # Wrap the model in FSDP.
+        log.info("Wrapping model with FSDP...")
+        assert cfg.fsdp is not None, "DistributedStrategy fsdp needs cfg.fsdp to be set!"
+        wrap_policy = olmo_model.get_fsdp_wrap_policy(cfg.fsdp.wrapping_strategy)
+
+        if version.parse(torch.__version__) >= version.parse("2.1.0"):
+            # This prevents any parameters from being initialized twice
+            def dummy_init_fn(module: torch.nn.Module) -> None:
+                module.to_empty(device=get_default_device())
+
+            param_init_fn = dummy_init_fn
+        else:
+            param_init_fn = None
+
+        # Set up device mesh for hybrid sharding in order to specify which nodes are assoicated to a given model replica
+        device_mesh = None
+        hybrid_sharding_fsdp_kwargs = {}
+        if cfg.fsdp.sharding_strategy in (ShardingStrategy.HYBRID_SHARD, ShardingStrategy._HYBRID_SHARD_ZERO2):
+            if version.parse(torch.__version__) < version.parse("2.2.0"):
+                # Device mesh was not added to PyTorch until v2.2.0
+                raise OLMoConfigurationError(
+                    "OLMo training does not correctly support hybrid sharding before torch 2.2.0"
+                )
+
+            from torch.distributed.device_mesh import init_device_mesh
+
+            num_model_replicas = cfg.fsdp.hybrid_sharding_num_model_replicas or (
+                get_world_size() // get_local_world_size()
+            )
+
+            if num_model_replicas <= 0:
+                raise OLMoConfigurationError("fsdp.hybrid_sharding_num_model_replicas must be a positive integer")
+
+            if get_world_size() % num_model_replicas != 0:
+                raise OLMoConfigurationError("fsdp.hybrid_sharding_num_model_replicas must divide world size")
+
+            device_mesh = init_device_mesh("cuda", (num_model_replicas, get_world_size() // num_model_replicas))
+            hybrid_sharding_fsdp_kwargs["device_mesh"] = device_mesh
+
+        dist_model = FSDP(
+            olmo_model,
+            sharding_strategy=cfg.fsdp.sharding_strategy,
+            mixed_precision=cfg.fsdp_precision,
+            auto_wrap_policy=wrap_policy,
+            use_orig_params=cfg.fsdp.use_orig_params,  # needed for compile and some of our optimizer/parameter metrics
+            limit_all_gathers=True,
+            device_id=get_local_rank(),
+            param_init_fn=param_init_fn,
+            **hybrid_sharding_fsdp_kwargs,
+        )
+    elif cfg.distributed_strategy is None:
+        raise NotImplementedError("Single accelerator training not implemented yet!")
+
     # when param_init_fn is None, FSDP will call reset_parameters() automatically
-    if param_init_fn is not None:
+    if param_init_fn is not None or cfg.distributed_strategy == DistributedStrategy.ddp:
         olmo_model.reset_parameters()
 
-    log.info(f"Peak GPU Memory (MB) after FSDP: {int(peak_gpu_memory() or 0)}")
+    log.info(f"Peak GPU Memory (MB) after {cfg.distributed_strategy}: {int(peak_gpu_memory() or 0)}")
     log.info("Model:")
-    log.info(fsdp_model)
+    log.info(dist_model)
 
     # Construct optimizer and learning rate scheduler.
-    optim = build_optimizer(cfg, fsdp_model)
+    optim = build_optimizer(cfg, dist_model)
     scheduler = build_scheduler(cfg)
 
     # Data indices file.
@@ -169,7 +230,7 @@ def main(cfg: TrainConfig) -> None:
         cfg=cfg,
         epoch=cfg.epoch,
         model=olmo_model,
-        fsdp_model=fsdp_model,
+        dist_model=dist_model,
         optim=optim,
         scheduler=scheduler,
         train_loader=train_loader,
@@ -178,9 +239,26 @@ def main(cfg: TrainConfig) -> None:
         indices_file=indices_file,
     ) as trainer:
         if not cfg.dry_run and not cfg.no_pre_train_checkpoint and cfg.load_path is None:
-            checkpoint_type = (
-                CheckpointType.sharded if cfg.save_num_checkpoints_to_keep != 0 else CheckpointType.unsharded
-            )
+            if cfg.distributed_strategy == DistributedStrategy.ddp:
+                checkpoint_type = CheckpointType.unsharded
+
+                if cfg.save_interval_unsharded is None:
+                    log.warning(
+                        "DDP requires setting `save_interval_unsharded`. Using the value set for `save_interval`."
+                    )
+                    cfg.save_interval_unsharded = cfg.save_interval
+
+                if cfg.save_num_unsharded_checkpoints_to_keep == 0:
+                    log.warning(
+                        "DDP requires setting `save_num_unsharded_checkpoints_to_keep`. Using the value set for `save_num_checkpoints_to_keep`."
+                    )
+                    cfg.save_num_unsharded_checkpoints_to_keep = cfg.save_num_checkpoints_to_keep
+            elif cfg.distributed_strategy == DistributedStrategy.fsdp:
+                checkpoint_type = (
+                    CheckpointType.sharded if cfg.save_num_checkpoints_to_keep != 0 else CheckpointType.unsharded
+                )
+            else:
+                raise NotImplementedError(f"Distributed strategy {cfg.distributed_strategy} not supported yet!")
 
             # We save a checkpoint up-front to make sure this won't fail (due to disk space or whatever).
             log.info("Saving pre-train checkpoint...")
@@ -217,7 +295,7 @@ def main(cfg: TrainConfig) -> None:
                     int(trainer.global_step + cfg.scheduler.t_warmup),
                 )
 
-        if cfg.force_save_unsharded:
+        if cfg.force_save_unsharded and cfg.distributed_strategy != DistributedStrategy.ddp:
             log.info("Saving unsharded checkpoint...")
             checkpoint_path, _ = trainer.save_checkpoint(checkpoint_type=CheckpointType.unsharded)
             log.info(f"Unsharded checkpoint saved to {checkpoint_path}")
@@ -246,13 +324,16 @@ if __name__ == "__main__":
         mp.set_start_method("spawn", force=True)
     except RuntimeError as e:
         print(f"failed to set multiprocessing start method: {e}")
+    log.info(f"Multiprocessing start method set to '{mp.get_start_method()}'")
 
     # Initialize process group.
     dist.init_process_group(backend="nccl")
+    log.info("Process group initialized")
 
     prepare_cli_environment()
+    log.info("CLI environment prepared")
 
-    log.info(f"multiprocessing start method set to '{mp.get_start_method()}'")
+    add_cached_path_clients()
 
     try:
         yaml_path, args_list = sys.argv[1], sys.argv[2:]